requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
idna==3.11
kiwisolver==1.4.9
kloppy==3.18.0
lxml==6.0.2
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sortedcontainers==2.4.0
tzdata==2025.2
yarl==1.22.0
